+++ lib/celluloid/events.rb
@@ -2,6 +2,35 @@ module Celluloid
   # High-priority internal system events
   class SystemEvent; end
 
+  # Request to link with another actor
+  class LinkingRequest < SystemEvent
+    attr_reader :actor, :type
+
+    def initialize(actor, type)
+      @actor, @type = actor, type.to_sym
+      raise ArgumentError, "type must be link or unlink" unless [:link, :unlink].include?(@type)
+    end
+
+    def process(links)
+      case type
+      when :link   then links << actor
+      when :unlink then links.delete actor
+      end
+
+      actor.mailbox << LinkingResponse.new(Actor.current, type)
+    end
+  end
+
+  # Response to a link request
+  class LinkingResponse
+    attr_reader :actor, :type
+
+    def initialize(actor, type)
+      @actor, @type = actor, type.to_sym
+      raise ArgumentError, "type must be link or unlink" unless [:link, :unlink].include?(@type)
+    end
+  end
+
   # An actor has exited for the given reason
   class ExitEvent < SystemEvent
     attr_reader :actor, :reason